[style] team: calculateProgress.ts 진행률 계산 로직 분리
